Make the events datetime method configurable
The method to retrieve an events date, was hardcoded to be "start_time".
If your event model did not have such a method, the advise was to simply
delegate "start_time" to whatever method returns a datetime obejct
representing the start time of your event model.

Instead of adjusting your model to be able to handle simple_calendar, we
should pass in the name of the method that should be called to retrieve
the start time through a configuration option.